+++ src/index.js
@@ -24,6 +24,10 @@ export { default as Pager } from '../stories/Components/Pager/Pager';
 export { default as FormAction } from '../stories/Components/Forms/FormAction/FormAction';
 export { default as TextCta } from '../stories/Components/TextCta/TextCta';
 export { default as Snackbar } from '../stories/Components/Snackbar/Snackbar';
+export {
+  default as UserFeedback,
+  DEFAULT_USER_FEEDBACK_LABELS,
+} from '../stories/Components/UserFeedback/UserFeedback';
 
 // Import global styles
 import '../stories/assets/scss/style.scss';